requirements.txt
Django==4.0
environs==9.3.5
whitenoise==5.3.0
dj-database-url==0.5.0
django-debug-toolbar==3.2.3
django-jazzmin==2.4.8
django-phonenumber-field==6.0.0
phonenumberslite==8.12.37
djangorestframework==3.13.1
drf-yasg==1.20.0
django-rq==2.5.1
